added handle altitude_real, precision(alti)
Added more info about the valve handle:

 - handle_altitude_real
 - handle_precision
 - handle_precisionalti

to the columns for the valve handle in order to avoid a possible problem.
Users could think that the altitude_real, precision and precisionalti columns refer to the handle or to the valve-node or both.
There are cases where the pipe location is not well determined, and the field workers could mistakenly think the pipe is right beneath.
In these cases, the actual pipe location can be even a meter away from the valve handle head that the workers see in the ground.
This could lead to more costly interventions if they are excavating to find the pipe as they may not find the pipe beneath which results in higher costs (time, more materials to cover the bigger hole etc.)
This commit permits to depict the situation where the valve node (situated on the pipe) is not directly beneath the valve handle (the valve is not in a vertical position).