[fix #93] avoid infinite recursion in geometry triggers
this avoids infinite recursion in geometry triggers when adding a pipe:
it was calling create node which was updating pipe and so on.

The triggers has been decoupled: the insertion of vertex to pipe is done AFTER node insertion/move

Original trigger was only adding vertex to pipe on node insertion, now it does also on upddate. @mhugo any objection?

One remaining question is regarding ST_Snap and 3D. Does it keep current Z values?