add hardcoded _schema_visible field to pipes and valves for enhanced perfomance on QGIS
This creates a field _schema_visible which determines if
an element is visible in schematic view.
It is calculated by COALESCE(schema_force_visible, value_list.schema_visible).
This allows direct use of this field in QGIS for rule-based labeling and
fixes performance issues if calculation is done on client side.

* schema_visible field is dropped on network_element as it is not used (was introduced for valves at the origin)
* _schema_visible and schema_force_visible are added to pipe and valve tables by the qwat_sys.fn_enable_schemavisible function
* old function qwat_sys.fn_enable_schemaview is enhanced and renamed